requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Django==5.1.2
djangorestframework==3.15.2
idna==3.10
lxml==5.3.0
prettytable==3.12.0
psycopg==3.2.3
pycparser==2.22
PyMuPDF==1.24.13
pytimeparse2==1.7.1
requests==2.32.3
sqlparse==0.5.1
sshkey-tools==0.11.3
urllib3==2.2.3
wcwidth==0.2.13
